polymorphism with interface doesnt work
If we are using derived class that implements an interface throught base type or that interface, then polymorphism doesnt work. It works only in case using derived type explicitly. "public new" it is not enough. The way to go is to explicitly implement the part of the interface that you want to override.
https://stackoverflow.com/questions/12314990/override-method-implementation-declared-in-an-interface